Refactor add and remove book functionality
